Put factories in the one class and added TodoBocks to Todo there.
